shopping cart

+++ src/main/java/com/app/JuberEats/entity/cart/ShoppingCartItem.java
@@ -8,9 +8,9 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ShoppingCartItem {
-    private Integer productId;
+    private Long productId;
     private String productName;
-    private Integer restaurantId;
+    private Long restaurantId;
     private String restaurantName;
     private Integer quantity;
 }
